Refactored EventController
