+++ app/Resources/views/adherent/contact.html.twig
@@ -11,7 +11,7 @@
                         {{ include('components/caret--left.html.twig') }}
                         Retour au comité</a>
                 {% else %}
-                    <a href="{{ path('app_committee_show_event', {'uuid': from.uuid.toString, 'slug': from.slug}) }}" class="back-to-list icon--with-text text--summary">
+                    <a href="{{ path('app_event_show', {'uuid': from.uuid.toString, 'slug': from.slug}) }}" class="back-to-list icon--with-text text--summary">
                         {{ include('components/caret--left.html.twig') }}
                         Retour à l'évenement</a>
                 {% endif %}
