+++ app/Resources/views/adherent/events.html.twig
@@ -46,7 +46,7 @@
                                 {% endif %}
                                 <h2 class="text--normal text--medium b__nudge--bottom-small text--breakword">
                                     <a class="link--no-decor"
-                                       href="{{ path('app_committee_show_event', { uuid: event.uuid, slug: event.slug }) }}">
+                                       href="{{ path('app_event_show', { uuid: event.uuid, slug: event.slug }) }}">
                                         {{ event.name }}
                                     </a>
                                 </h2>
@@ -77,7 +77,7 @@
                                     </div>
                                 {% endif %}
 
-                                <a href="{{ path('app_committee_show_event', { uuid: event.uuid, slug: event.slug }) }}"
+                                <a href="{{ path('app_event_show', { uuid: event.uuid, slug: event.slug }) }}"
                                    class="btn" target="_blank" rel="noopener noreferrer">
                                     En savoir plus
                                 </a>
@@ -93,7 +93,7 @@
                              title="Carte d'accès à l'événement {{ event.name }}"
                              alt="{{ event.inlineFormattedAddress }}" />
 
-                        <a href="{{ path('app_committee_show_event', { uuid: event.uuid, slug: event.slug }) }}"
+                        <a href="{{ path('app_event_show', { uuid: event.uuid, slug: event.slug }) }}"
                            class="search__results__row__link link--no-decor" target="_blank" rel="noopener noreferrer">
                             En savoir plus
                         </a>
