+++ app/Resources/views/admin/event_actions.html.twig
@@ -13,7 +13,7 @@
         </a>
     </div>
     <div style="width: 160px; margin-top: 5px;">
-        <a href="{{ path('app_committee_show_event', {'uuid': object.uuid.toString, 'slug': object.slug}) }}"
+        <a href="{{ path('app_event_show', {'uuid': object.uuid.toString, 'slug': object.slug}) }}"
            class="btn btn-xs btn-default" title="Voir sur le site" target="_blank">
             <i class="fa fa-eye" aria-hidden="true"></i>
             Voir sur le site
