+++ app/Resources/views/events/_header.html.twig
@@ -26,7 +26,7 @@
                 L'événement à été annulé
             </div>
         {% elseif not event.full %}
-            <a class="btn btn--no-border btn--full" href="{{ path('app_committee_attend_event', {uuid: event.uuid.toString, slug: event.slug}) }}">Je veux participer</a>
+            <a class="btn btn--no-border btn--full" href="{{ path('app_event_attend', {uuid: event.uuid.toString, slug: event.slug}) }}">Je veux participer</a>
         {% else %}
             <button class="btn btn--no-border btn--full btn--text-blue btn--disabled" disabled="disabled">
                 Je veux participer
@@ -54,7 +54,7 @@
                 <i class="fa fa-envelope-o"></i>
             </li>
             <li role="button" aria-label="{{ 'committee.event.show.share.download_ical'|trans }}">
-                <a href="{{ path('app_committee_event_export_ical', {uuid: event.uuid, slug: event.slug}) }}" title="{{ 'committee.event.show.share.download_ical'|trans }}" class="text--white">
+                <a href="{{ path('app_event_export_ical', {uuid: event.uuid, slug: event.slug}) }}" title="{{ 'committee.event.show.share.download_ical'|trans }}" class="text--white">
                     <i class="fa fa-calendar-plus-o" aria-hidden="true"></i>
                     <span class="sr-only">{{ 'committee.event.show.share.download_ical'|trans }}</span>
                 </a>
