+++ app/Resources/views/events/attend.html.twig
@@ -5,7 +5,7 @@
 {% block content %}
 
     <section class="l__wrapper">
-        <a href="{{ path('app_committee_show_event', {'uuid': committee_event.uuid.toString, 'slug': committee_event.slug}) }}"
+        <a href="{{ path('app_event_show', {'uuid': committee_event.uuid.toString, 'slug': committee_event.slug}) }}"
            class="previous-link icon--with-text text--summary">
             {{ include('components/caret--left.html.twig') }}
             Retour à la page de l'événement
@@ -25,7 +25,7 @@
                 </p>
             {% endif %}
 
-            <form action="{{ path('app_committee_attend_event', {'uuid': committee_event.uuid.toString, 'slug': committee_event.slug}) }}" method="post" id="committee-event-registration-form">
+            <form action="{{ path('app_event_attend', {'uuid': committee_event.uuid.toString, 'slug': committee_event.slug}) }}" method="post" id="committee-event-registration-form">
                 {{ form_errors(form) }}
 
                 <div class="form__row" id="field-first-name">
@@ -71,7 +71,7 @@
                     <button type="submit" class="btn btn--blue" id="attend-committee-event">
                         Je m'inscris
                     </button>
-                    <a  href="{{ path('app_committee_show_event', {'uuid': committee_event.uuid.toString, 'slug': committee_event.slug}) }}"
+                    <a  href="{{ path('app_event_show', {'uuid': committee_event.uuid.toString, 'slug': committee_event.slug}) }}"
                         class="btn">
                         Annuler
                     </a>
