+++ app/Resources/views/events/cancel.html.twig
@@ -3,7 +3,7 @@
 {% block page_title event.name %}
 
 {% block content %}
-    {% set event_url = url('app_committee_show_event', {uuid: event.uuid.toString, slug: event.slug}) %}
+    {% set event_url = url('app_event_show', {uuid: event.uuid.toString, slug: event.slug}) %}
     <main class="committee__event__create">
         <section class="commitee__event l__wrapper">
             <div class="committee__event__header">
@@ -22,7 +22,7 @@
                     </div>
                 </div>
                 <div class="committee__event__header__cta">
-                    <a class="btn btn--no-border btn--full" href="{{ path('app_committee_attend_event', {uuid: event.uuid.toString, slug: event.slug}) }}">Je veux participer</a>
+                    <a class="btn btn--no-border btn--full" href="{{ path('app_event_attend', {uuid: event.uuid.toString, slug: event.slug}) }}">Je veux participer</a>
                     <div class="text--body text--white icon--with-text l__row--center b__nudge--top-10">
                         {{ include('components/participants.html.twig') }}
                         {{ 'event.participants_count'|transchoice(event.participantsCount) }}
@@ -47,7 +47,7 @@
                     </ul>
                 </div>
             </div>
-            <a href="{{ path('app_committee_show_event', {'uuid': event.uuid.toString, 'slug': event.slug}) }}" class="back-to-list icon--with-text b__nudge--top-15">
+            <a href="{{ path('app_event_show', {'uuid': event.uuid.toString, 'slug': event.slug}) }}" class="back-to-list icon--with-text b__nudge--top-15">
                 {{ include('components/caret--left.html.twig') }}
                 Retour à l'événement
             </a>
@@ -72,7 +72,7 @@
                         <button type="submit" class="btn btn--blue">
                             Oui, annuler l'événement
                         </button>
-                        <a href="{{ path('app_committee_show_event', {'uuid': event.uuid.toString, 'slug': event.slug}) }}"
+                        <a href="{{ path('app_event_show', {'uuid': event.uuid.toString, 'slug': event.slug}) }}"
                            class="btn">
                             Non, retour
                         </a>
