+++ app/Resources/views/events/invitation_sent.html.twig
@@ -4,7 +4,7 @@
 
 {% block content %}
     <section class="l__wrapper">
-        <a href="{{ path('app_committee_show_event', {'uuid': committee_event.uuid, 'slug': committee_event.slug}) }}"
+        <a href="{{ path('app_event_show', {'uuid': committee_event.uuid, 'slug': committee_event.slug}) }}"
            class="previous-link icon--with-text text--summary">
             {{ include('components/caret--left.html.twig') }}
             Retour à la page de l'événement
