+++ app/Resources/views/events/members.html.twig
@@ -91,7 +91,7 @@
 {% endblock %}
 
 {% block content %}
-    {% set event_url = url('app_committee_show_event', {uuid: event.uuid.toString, slug: event.slug}) %}
+    {% set event_url = url('app_event_show', {uuid: event.uuid.toString, slug: event.slug}) %}
     <main class="committee__event__create">
         <section class="commitee__event l__wrapper">
             <div class="committee__event__header">
@@ -118,7 +118,7 @@
                             L'événement à été annulé
                         </div>
                     {% else %}
-                        <a class="btn btn--no-border btn--full" href="{{ path('app_committee_attend_event', {uuid: event.uuid.toString, slug: event.slug}) }}">Je veux participer</a>
+                        <a class="btn btn--no-border btn--full" href="{{ path('app_event_attend', {uuid: event.uuid.toString, slug: event.slug}) }}">Je veux participer</a>
                     {% endif %}
                     <div class="text--body text--white icon--with-text l__row--center b__nudge--top-10">
                         {{ include('components/participants.html.twig') }}
@@ -144,7 +144,7 @@
                     </ul>
                 </div>
             </div>
-            <a href="{{ path('app_committee_show_event', {'uuid': event.uuid.toString, 'slug': event.slug}) }}" class="back-to-list icon--with-text b__nudge--top-15">
+            <a href="{{ path('app_event_show', {'uuid': event.uuid.toString, 'slug': event.slug}) }}" class="back-to-list icon--with-text b__nudge--top-15">
                 {{ include('components/caret--left.html.twig') }}
                 Retour à l'événement
             </a>
