+++ app/Resources/views/events/show.html.twig
@@ -1,7 +1,7 @@
 {% extends 'base.html.twig' %}
 
 {% block page_title event.name %}
-{% block canonical_url url('app_committee_show_event', {'uuid': event.uuid.toString, 'slug': event.slug}) %}
+{% block canonical_url url('app_event_show', {'uuid': event.uuid.toString, 'slug': event.slug}) %}
 
 {% block json_ld %}
     <script type="application/ld+json">
@@ -73,7 +73,7 @@
 {% endblock %}
 
 {% block content %}
-    {% set event_url = url('app_committee_show_event', {uuid: event.uuid.toString, slug: event.slug}) %}
+    {% set event_url = url('app_event_show', {uuid: event.uuid.toString, slug: event.slug}) %}
     <main>
         <section class="commitee__event l__wrapper">
             <a href="{{ path('app_search_events') }}" class="back-to-list icon--with-text text--summary">
@@ -129,7 +129,7 @@
                         </li>
                     {% endif %}
                     <li>
-                        <a href="{{ path('app_event_registrations', {'uuid': event.uuid.toString, 'slug': event.slug}) }}">
+                        <a href="{{ path('app_event_members', {'uuid': event.uuid.toString, 'slug': event.slug}) }}">
                             Gérer les participants
                         </a>
                     </li>
