+++ src/Controller/EnMarche/AdherentController.php
@@ -209,7 +209,7 @@ class AdherentController extends Controller
             }
 
             if ($from instanceof Event) {
-                return $this->redirectToRoute('app_committee_show_event', [
+                return $this->redirectToRoute('app_event_show', [
                     'uuid' => $from->getUuid()->toString(),
                     'slug' => $from->getSlug(),
                 ]);
