+++ src/Controller/EnMarche/CommitteeManagerController.php
@@ -74,7 +74,7 @@ class CommitteeManagerController extends Controller
 
             $this->addFlash('info', $this->get('translator')->trans('committee.event.creation.success'));
 
-            return $this->redirectToRoute('app_committee_show_event', [
+            return $this->redirectToRoute('app_event_show', [
                 'uuid' => (string) $command->getEvent()->getUuid(),
                 'slug' => (string) $command->getEvent()->getSlug(),
             ]);
