+++ src/Referent/ManagedEventsExporter.php
@@ -28,7 +28,7 @@ class ManagedEventsExporter
                 'id' => $event->getId(),
                 'name' => [
                     'label' => $event->getName(),
-                    'url' => $this->urlGenerator->generate('app_committee_show_event', [
+                    'url' => $this->urlGenerator->generate('app_event_show', [
                         'uuid' => $event->getUuid()->toString(),
                         'slug' => $event->getSlug(),
                     ]),
