+++ tests/Controller/EnMarche/EventControllerTest.php
@@ -8,12 +8,9 @@ use AppBundle\DataFixtures\ORM\LoadEventData;
 use AppBundle\DataFixtures\ORM\LoadHomeBlockData;
 use AppBundle\Entity\EventInvite;
 use AppBundle\Entity\EventRegistration;
-use AppBundle\Mailjet\Message\EventCancellationMessage;
-use AppBundle\Mailjet\Message\EventContactMembersMessage;
 use AppBundle\Mailjet\Message\EventInvitationMessage;
 use AppBundle\Mailjet\Message\EventRegistrationConfirmationMessage;
 use AppBundle\Repository\EventRegistrationRepository;
-use Ramsey\Uuid\Uuid;
 use Symfony\Component\DomCrawler\Crawler;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -160,288 +157,6 @@ class EventControllerTest extends MysqlWebTestCase
 
         $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
         $this->assertContains("L'événement est complet.", $crawler->filter('.form__errors')->text());
-    }
-
-    /**
-     * @dataProvider provideHostProtectedPages
-     */
-    public function testAnonymousUserCannotEditEvent($path)
-    {
-        $this->client->request('GET', $path);
-        $this->assertResponseStatusCode(Response::HTTP_FOUND, $this->client->getResponse());
-    }
-
-    /**
-     * @dataProvider provideCancelledInaccessiblePages
-     */
-    public function testRegisteredAdherentUserCannotFoundPagesOfCancelledEvent($path)
-    {
-        $this->authenticateAsAdherent($this->client, 'benjyd@aol.com', 'HipHipHip');
-        $this->client->request(Request::METHOD_GET, $path);
-        $this->assertResponseStatusCode(Response::HTTP_NOT_FOUND, $this->client->getResponse());
-    }
-
-    /**
-     * @dataProvider provideHostProtectedPages
-     */
-    public function testRegisteredAdherentUserCannotEditEvent($path)
-    {
-        $this->authenticateAsAdherent($this->client, 'benjyd@aol.com', 'HipHipHip');
-        $this->client->request('GET', $path);
-        $this->assertResponseStatusCode(Response::HTTP_FORBIDDEN, $this->client->getResponse());
-    }
-
-    public function provideHostProtectedPages()
-    {
-        $uuid = LoadEventData::EVENT_1_UUID;
-        $slug = date('Y-m-d', strtotime('+3 days')).'-reunion-de-reflexion-parisienne';
-
-        return [
-            ['/evenements/'.$uuid.'/'.$slug.'/modifier'],
-            ['/evenements/'.$uuid.'/'.$slug.'/inscrits'],
-        ];
-    }
-
-    public function provideCancelledInaccessiblePages()
-    {
-        $uuid = LoadEventData::EVENT_6_UUID;
-        $slug = date('Y-m-d', strtotime('+60 days')).'-reunion-de-reflexion-parisienne-annule';
-
-        return [
-            ['/evenements/'.$uuid.'/'.$slug.'/modifier'],
-            ['/evenements/'.$uuid.'/'.$slug.'/inscription'],
-            ['/evenements/'.$uuid.'/'.$slug.'/annuler'],
-        ];
-    }
-
-    public function testOrganizerCanEditEvent()
-    {
-        $this->authenticateAsAdherent($this->client, 'jacques.picard@en-marche.fr', 'changeme1337');
-
-        $crawler = $this->client->request('GET', '/evenements/'.LoadEventData::EVENT_1_UUID.'/'.date('Y-m-d', strtotime('+3 days')).'-reunion-de-reflexion-parisienne/modifier');
-
-        $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
-
-        $this->client->submit($crawler->selectButton('Enregistrer')->form([
-            'committee_event' => [
-                'name' => "Débat sur l'écologie",
-                'description' => 'Cette journée sera consacrée à un grand débat sur la question écologique.',
-                'category' => $this->getEventCategoryIdForName(LoadEventCategoryData::LEGACY_EVENT_CATEGORIES['CE003']),
-                'address' => [
-                    'address' => '6 rue Neyret',
-                    'country' => 'FR',
-                    'postalCode' => '69001',
-                    'city' => '69001-69381',
-                    'cityName' => '',
-                ],
-                'beginAt' => [
-                    'date' => [
-                        'year' => '2022',
-                        'month' => '3',
-                        'day' => '2',
-                    ],
-                    'time' => [
-                        'hour' => '9',
-                        'minute' => '30',
-                    ],
-                ],
-                'finishAt' => [
-                    'date' => [
-                        'year' => '2022',
-                        'month' => '3',
-                        'day' => '2',
-                    ],
-                    'time' => [
-                        'hour' => '19',
-                        'minute' => '0',
-                    ],
-                ],
-                'capacity' => '1500',
-            ],
-        ]));
-
-        $this->assertStatusCode(Response::HTTP_FOUND, $this->client);
-
-        // Follow the redirect and check the adherent can see the committee page
-        $crawler = $this->client->followRedirect();
-        $this->assertStatusCode(Response::HTTP_OK, $this->client);
-        $this->assertContains('L\'événement a bien été modifié.', $crawler->filter('#notice-flashes')->text());
-        $this->assertSame('Débat sur l\'écologie | En Marche !', $crawler->filter('title')->text());
-        $this->assertSame('Débat sur l\'écologie', $crawler->filter('.committee-event-name')->text());
-        $this->assertSame('Organisé par Jacques Picard du comité En Marche Paris 8', trim(preg_replace('/\s+/', ' ', $crawler->filter('.committee-event-organizer')->text())));
-        $this->assertSame('Mercredi 2 mars 2022, 9h30', $crawler->filter('.committee-event-date')->text());
-        $this->assertSame('6 rue Neyret, 69001 Lyon 1er', $crawler->filter('.committee-event-address')->text());
-        $this->assertSame('Cette journée sera consacrée à un grand débat sur la question écologique.', $crawler->filter('.committee-event-description')->text());
-    }
-
-    public function testOrganizerCanCancelEvent()
-    {
-        $this->authenticateAsAdherent($this->client, 'francis.brioul@yahoo.com', 'Champion20');
-
-        $crawler = $this->client->request(Request::METHOD_GET, '/evenements/'.LoadEventData::EVENT_2_UUID.'/'.date('Y-m-d', strtotime('+10 days')).'-reunion-de-reflexion-dammarienne/annuler');
-
-        $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
-
-        $this->client->submit($crawler->selectButton('Oui, annuler l\'événement')->form());
-
-        $this->assertStatusCode(Response::HTTP_FOUND, $this->client);
-
-        // Follow the redirect and check the adherent can see the committee page
-        $crawler = $this->client->followRedirect();
-        $this->assertStatusCode(Response::HTTP_OK, $this->client);
-        static::assertContains('L\'événement a bien été annulé.', $crawler->filter('#notice-flashes')->text());
-
-        $messages = $this->getMailjetEmailRepository()->findMessages(EventCancellationMessage::class);
-        /** @var EventCancellationMessage $message */
-        $message = array_shift($messages);
-
-        // Two mails have been sent
-        static::assertCount(2, $message->getRecipients());
-    }
-
-    public function testCommitteeHostCanEditEvent()
-    {
-        $this->authenticateAsAdherent($this->client, 'gisele-berthoux@caramail.com', 'ILoveYouManu');
-
-        $this->client->request('GET', '/evenements/'.LoadEventData::EVENT_1_UUID.'/'.date('Y-m-d', strtotime('+3 days')).'-reunion-de-reflexion-parisienne/modifier');
-
-        $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
-    }
-
-    public function testOrganizerCanSeeRegistrations()
-    {
-        $this->authenticateAsAdherent($this->client, 'jacques.picard@en-marche.fr', 'changeme1337');
-        $crawler = $this->client->request('GET', '/evenements/'.LoadEventData::EVENT_1_UUID.'/'.date('Y-m-d', strtotime('+3 days')).'-reunion-de-reflexion-parisienne');
-        $crawler = $this->client->click($crawler->selectLink('Gérer les participants')->link());
-
-        $this->assertTrue($this->seeMembersList($crawler, 2));
-    }
-
-    public function testOrganizerCanExportRegistrationsWithWrongUuids()
-    {
-        $this->authenticateAsAdherent($this->client, 'jacques.picard@en-marche.fr', 'changeme1337');
-
-        $crawler = $this->client->request('GET', '/evenements/'.LoadEventData::EVENT_1_UUID.'/'.date('Y-m-d', strtotime('+3 days')).'-reunion-de-reflexion-parisienne');
-        $crawler = $this->client->click($crawler->selectLink('Gérer les participants')->link());
-
-        $exportUrl = $this->client->getRequest()->getPathInfo().'/exporter';
-
-        $this->client->request(Request::METHOD_POST, $exportUrl, [
-            'token' => $crawler->filter('#members-export-token')->attr('value'),
-            'exports' => json_encode(['wrong_uuid']),
-        ]);
-
-        $this->assertStatusCode(Response::HTTP_BAD_REQUEST, $this->client);
-    }
-
-    public function testOrganizerCanExportRegistrations()
-    {
-        $this->authenticateAsAdherent($this->client, 'jacques.picard@en-marche.fr', 'changeme1337');
-        $crawler = $this->client->request('GET', '/evenements/'.LoadEventData::EVENT_1_UUID.'/'.date('Y-m-d', strtotime('+3 days')).'-reunion-de-reflexion-parisienne');
-        $crawler = $this->client->click($crawler->selectLink('Gérer les participants')->link());
-
-        $token = $crawler->filter('#members-export-token')->attr('value');
-        $uuids = (array) $crawler->filter('input[name="members[]"]')->attr('value');
-
-        $exportUrl = $this->client->getRequest()->getPathInfo().'/exporter';
-
-        $this->client->request(Request::METHOD_POST, $exportUrl, [
-            'token' => $token,
-            'exports' => json_encode($uuids),
-        ]);
-
-        $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
-        $this->assertCount(3, explode("\n", $this->client->getResponse()->getContent()));
-
-        // Try to illegally export an adherent data
-        $uuids[] = Uuid::uuid4();
-
-        $this->client->request(Request::METHOD_POST, $exportUrl, [
-            'token' => $token,
-            'exports' => json_encode($uuids),
-        ]);
-
-        $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
-        $this->assertCount(3, explode("\n", $this->client->getResponse()->getContent()));
-
-        $this->client->request(Request::METHOD_POST, $exportUrl, [
-            'token' => $token,
-            'exports' => json_encode([]),
-        ]);
-
-        $this->assertResponseStatusCode(Response::HTTP_FOUND, $this->client->getResponse());
-    }
-
-    public function testOrganizerCanContactRegistrations()
-    {
-        $this->authenticateAsAdherent($this->client, 'jacques.picard@en-marche.fr', 'changeme1337');
-        $crawler = $this->client->request('GET', '/evenements/'.LoadEventData::EVENT_1_UUID.'/'.date('Y-m-d', strtotime('+3 days')).'-reunion-de-reflexion-parisienne');
-        $crawler = $this->client->click($crawler->selectLink('Gérer les participants')->link());
-
-        $token = $crawler->filter('#members-contact-token')->attr('value');
-        $uuids = (array) $crawler->filter('input[name="members[]"]')->attr('value');
-
-        $membersUrl = $this->client->getRequest()->getPathInfo();
-        $contactUrl = $membersUrl.'/contacter';
-
-        $crawler = $this->client->request(Request::METHOD_POST, $contactUrl, [
-            'token' => $token,
-            'contacts' => json_encode($uuids),
-        ]);
-
-        $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
-
-        // Try to post with an empty message
-        $crawler = $this->client->request(Request::METHOD_POST, $contactUrl, [
-            'token' => $crawler->filter('input[name="token"]')->attr('value'),
-            'contacts' => $crawler->filter('input[name="contacts"]')->attr('value'),
-            'message' => ' ',
-        ]);
-
-        $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
-        $this->assertSame('Cette valeur ne doit pas être vide.', $crawler->filter('.form__errors > .form__error')->text());
-
-        $this->client->request(Request::METHOD_POST, $contactUrl, [
-            'token' => $crawler->filter('input[name="token"]')->attr('value'),
-            'contacts' => $crawler->filter('input[name="contacts"]')->attr('value'),
-            'message' => 'Hello!',
-        ]);
-
-        $this->assertClientIsRedirectedTo($membersUrl, $this->client);
-
-        $crawler = $this->client->followRedirect();
-
-        $this->seeMessageSuccesfullyCreatedFlash($crawler, 'Félicitations, votre message a bien été envoyé aux inscrits sélectionnés.');
-
-        // Email should have been sent
-        $this->assertCount(1, $this->getMailjetEmailRepository()->findMessages(EventContactMembersMessage::class));
-
-        // Try to illegally contact an adherent
-        $uuids[] = Uuid::uuid4();
-
-        $crawler = $this->client->request(Request::METHOD_POST, $contactUrl, [
-            'token' => $token,
-            'contacts' => json_encode($uuids),
-        ]);
-
-        $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
-        $this->assertCount(1, json_decode($crawler->filter('input[name="contacts"]')->attr('value'), true));
-
-        // Force the contact form with foreign uuid
-        $this->client->request(Request::METHOD_POST, $contactUrl, [
-            'token' => $crawler->filter('input[name="token"]')->attr('value'),
-            'contacts' => json_encode($uuids),
-            'message' => 'Hello!',
-        ]);
-
-        $this->assertClientIsRedirectedTo($membersUrl, $this->client);
-    }
-
-    public function testExportIcalEvent()
-    {
-        $this->client->request('GET', '/evenements/'.LoadEventData::EVENT_1_UUID.'/'.date('Y-m-d', strtotime('+3 days')).'-reunion-de-reflexion-parisienne/ical');
-
-        $this->assertResponseStatusCode(Response::HTTP_OK, $this->client->getResponse());
     }
 
     public function testAdherentCanInviteToEvent()
@@ -598,12 +313,6 @@ class EventControllerTest extends MysqlWebTestCase
         }
 
         return 1 === count($flash);
-    }
-
-    private function seeMembersList(Crawler $crawler, int $count): bool
-    {
-        // Header row is part of the count
-        return $count === count($crawler->filter('table > tr'));
     }
 
     protected function setUp()
